Moved hoc folder in components folder
