+++ src/components/Profile/ProfileContainer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import About from './About'
 import Profile from './Profile'
 import FriendsList from './FriendsList'
-import authHoc from '../../hoc/authHoc'
+import authHoc from '../hoc/authHoc'
 import profileStore from '../../mobx/profile'
 import friendsStore from '../../mobx/friends'
 import { type ProfileData } from '../../types'
